+++ lib/spells/effects/UnitEffect.cpp
@@ -223,7 +223,7 @@ EffectTarget UnitEffect::transformTargetByChain(const Mechanics * m, const Targe
 			effectTarget.emplace_back();
 
 		for(auto hex : battle::Unit::getHexes(unit->getPosition(), unit->doubleWide(), unit->unitSide()))
-			possibleHexes.erase(hex);
+			possibleHexes.erase(hex.toInt());
 
 		if(possibleHexes.empty())
 			break;
